feat: Update lambda to rename log as yesterday #252

+++ server_Lambda/index.js
@@ -3,7 +3,11 @@ console.log("Loading function");
 const aws = require("aws-sdk");
 const strftime = require("strftime");
 const strftimeKOR = strftime.timezone("+0900");
-const date = strftimeKOR("%F", new Date());
+
+// set date as yesterday
+const _date = new Date();
+_date.setDate(_date.getDate() - 1);
+const date = strftimeKOR("%F", _date);
 
 exports.handler = async (event, context) => {
   console.log("Received event:", JSON.stringify(event, null, 2));
@@ -31,20 +35,23 @@ exports.handler = async (event, context) => {
     for (let content of data.Contents) {
       const keys = content.Key.split("/");
       if (keys.length === 2 && keys[1] !== "") {
-        await s3
-          .copyObject({
-            Bucket: bucket,
-            CopySource: `${bucket}/${content.Key}`, // old file Key
-            Key: `archived/${date}/${content.Key.replace("raw/", "")}`, // new file Key
-          })
-          .promise();
+        const subKeys = keys[1].split(" ");
+        if (subKeys.length === 3) {
+          await s3
+            .copyObject({
+              Bucket: bucket,
+              CopySource: `${bucket}/${content.Key}`, // old file Key
+              Key: `archived/${date}/${date} ${subKeys[2]}`, // new file Key
+            })
+            .promise();
 
-        await s3
-          .deleteObject({
-            Bucket: bucket,
-            Key: content.Key,
-          })
-          .promise();
+          await s3
+            .deleteObject({
+              Bucket: bucket,
+              Key: content.Key,
+            })
+            .promise();
+        }
       }
     }
